move brave and bitgo controllers into main src tree
