+++ app/scripts/controllers/bitgo.js
