+++ brave/ui/app/pages/connect-accounts/connect-accounts.component.js
@@ -10,7 +10,7 @@ import {
   BRAVE_BITGO_WALLET_INDEX
 } from '../../helpers/constants/routes'
 
-const { supportedCoins } = require('../../../../../brave/app/scripts/controllers/bitgo')
+const { supportedCoins } = require('../../../../../app/scripts/controllers/bitgo')
 
 module.exports = class BraveConnectAccounts extends PureComponent {
   static contextTypes = {
